+++ Trarizon.Library/Collections/Extensions/EnumerableQueries/Single.cs
@@ -3,30 +3,44 @@
 namespace Trarizon.Library.Collections.Extensions;
 partial class EnumerableQuery
 {
-    // TODO?: when defaultValue is not null, returns false, there's still warning
-
     /// <summary>
     /// Try get first element in sequence,
     /// this method returns false when there is not exactly one element in sequence
     /// </summary>
+    /// <remarks>
+    /// About nullable analysis:<br/>
+    /// There's no warning if <typeparamref name="T"/> is nullable reference type.<br/>
+    /// It's hard to provide perfect in all conditions,
+    /// so I chose a least-used condition (IMO). At least
+    /// it provide correct analysis in most conditions where
+    /// <typeparamref name="T"/> is not null
+    /// </remarks>
     /// <param name="value">
     /// The only element in sequence,
     /// or <paramref name="defaultValue"/> if sequence is empty or contains more than 1 element.
     /// </param>
     /// <returns><see langword="true"/> if there is exactly one elements in sequence</returns>
-    public static bool TrySingle<T>(this IEnumerable<T> source, [MaybeNullWhen(false)] out T value, T? defaultValue = default!)
+    public static bool TrySingle<T>(this IEnumerable<T> source, [NotNullWhen(true), NotNullIfNotNull(nameof(defaultValue))] out T? value, T? defaultValue = default!)
         => source.TrySingleInternal(out value, defaultValue, false);
 
     /// <summary>
     /// Try get first element satisfying specific condition in sequence,
     /// this method returns false when there is not exactly one element satisfying condition in sequence
     /// </summary>
+    /// <remarks>
+    /// About nullable analysis:<br/>
+    /// There's no warning if <typeparamref name="T"/> is nullable reference type.<br/>
+    /// It's hard to provide perfect in all conditions,
+    /// so I chose a least-used condition (IMO). At least
+    /// it provide correct analysis in most conditions where
+    /// <typeparamref name="T"/> is not null
+    /// </remarks>
     /// <param name="value">
     /// The only element in sequence,
     /// or <paramref name="defaultValue"/> if sequence is empty or contains more than 1 element.
     /// </param>
     /// <returns><see langword="true"/> if there is exactly one elements satisfying condition in sequence</returns>
-    public static bool TrySingle<T>(this IEnumerable<T> source, Func<T, bool> predicate, [MaybeNullWhen(false)] out T value, T? defaultValue = default!)
+    public static bool TrySingle<T>(this IEnumerable<T> source, Func<T, bool> predicate, [NotNullWhen(true), NotNullIfNotNull(nameof(defaultValue))] out T? value, T? defaultValue = default!)
         => source.TryPredicateSingleInternal(predicate, out value, defaultValue, false);
 
     /// <summary>
@@ -38,7 +52,7 @@ partial class EnumerableQuery
     /// or <paramref name="defaultValue"/> if sequence is empty or contains more than 1 element.
     /// </param>
     /// <returns><see langword="true"/> if there is one or zero elements in sequence</returns>
-    public static bool TrySingleOrNone<T>(this IEnumerable<T> source, [MaybeNullWhen(false)] out T value, T? defaultValue = default!)
+    public static bool TrySingleOrNone<T>(this IEnumerable<T> source, [NotNullIfNotNull(nameof(defaultValue))] out T? value, T? defaultValue = default!)
         => source.TrySingleInternal(out value, defaultValue, true);
 
     /// <summary>
@@ -50,10 +64,10 @@ partial class EnumerableQuery
     /// or <paramref name="defaultValue"/> if sequence is empty or contains more than 1 element.
     /// </param>
     /// <returns><see langword="true"/> if there is one or zero elements satisfying condition in sequence</returns>
-    public static bool TrySingleOrNone<T>(this IEnumerable<T> source, Func<T, bool> predicate, [MaybeNullWhen(false)] out T value, T? defaultValue = default!)
+    public static bool TrySingleOrNone<T>(this IEnumerable<T> source, Func<T, bool> predicate, [NotNullIfNotNull(nameof(defaultValue))] out T? value, T? defaultValue = default!)
         => source.TryPredicateSingleInternal(predicate, out value, defaultValue, true);
 
-    private static bool TrySingleInternal<T>(this IEnumerable<T> source, [MaybeNullWhen(false)] out T value, T? defaultValue, bool resultWhenZero)
+    private static bool TrySingleInternal<T>(this IEnumerable<T> source, out T? value, T? defaultValue, bool resultWhenZero)
     {
         if (source is IList<T> list) {
             switch (list.Count) {
@@ -86,7 +100,7 @@ partial class EnumerableQuery
         return false;
     }
 
-    private static bool TryPredicateSingleInternal<T>(this IEnumerable<T> source, Func<T, bool> predicate, [MaybeNullWhen(false)] out T value, T? defaultValue, bool resultWhenNotFound)
+    private static bool TryPredicateSingleInternal<T>(this IEnumerable<T> source, Func<T, bool> predicate, out T? value, T? defaultValue, bool resultWhenNotFound)
     {
         using var enumerator = source.GetEnumerator();